+++ RSO/txredisapi/model.py
@@ -1,11 +1,9 @@
-from typing import Union
+from typing import List, Optional, Union
 
-from dataclasses import asdict
 from txredisapi import BaseRedisProtocol, ConnectionHandler
 from twisted.internet.defer import inlineCallbacks
 
 from RSO.base import BaseModel
-from .index import ListIndex
 
 
 class Model(BaseModel):
@@ -33,7 +31,7 @@ class Model(BaseModel):
 
     @classmethod
     @inlineCallbacks
-    def search(cls, redis: ConnectionHandler, value):
+    def search(cls, redis: ConnectionHandler, value) -> Optional['Model']:
         redis_key = cls.redis_key_from_value(value)
         result = yield redis.exists(redis_key)
         if bool(result):
@@ -42,6 +40,33 @@ class Model(BaseModel):
             dict_data = cls.from_redis(dict(zip(fields, redis_data)))
             return cls(**dict_data)
         return
+
+    @classmethod
+    @inlineCallbacks
+    def all(cls, redis: ConnectionHandler) -> List['Model']:
+        redis_key = cls.redis_key_from_value('*')
+        members = yield redis.keys(redis_key)
+        indexes = []
+        for index_class in cls.__indexes__:
+            indexes.append(
+                f'::{index_class.__index_name__}::{index_class.__key__}'
+            )
+
+        pipe = yield redis.multi()
+        for member in members:
+            is_index = False
+            for index in indexes:
+                if index in member:
+                    is_index = True
+                    break
+            if not is_index:
+                yield pipe.hgetall(member)
+        result_data = yield pipe.commit()
+
+        result = []
+        for data in result_data:
+            result.append(cls(**data))
+        return result
 
     @inlineCallbacks
     def delete(self, redis: Union[BaseRedisProtocol, ConnectionHandler]):
@@ -56,48 +81,6 @@ class Model(BaseModel):
             if getattr(self, index_class.__key__) is None:
                 continue
             index_class.remove(pipe, self)
-        pipe.delete(self.redis_key)
+        yield pipe.delete(self.redis_key)
         if do_commit:
             yield pipe.commit()
-
-    # @inlineCallbacks
-    # def extended_save(
-    #     self, redis: Union[BaseRedisProtocol, ConnectionHandler]
-    # ):
-    #     """Extended save function to avoid multiple push on list index
-    #
-    #     Note: Use this method if you use List Index
-    #     """
-    #     list_index_map = {}
-    #     for index_class in self.__indexes__ or []:
-    #         if not issubclass(index_class, ListIndex):
-    #             continue
-    #         index = index_class.create_from_model_class(self)
-    #         model_key_value = getattr(self, self.__key__, None)
-    #         if model_key_value is None:
-    #             exist_on_index = False
-    #         else:
-    #             exist_on_index = yield index.is_exist_on_list(
-    #                 redis, model_key_value
-    #             )
-    #         list_index_map[index] = exist_on_index
-    #
-    #     if isinstance(redis, ConnectionHandler):
-    #         pipe = yield redis.multi()
-    #     else:
-    #         raise NotImplementedError
-    #
-    #     pipe.hmset(self.redis_key, self.to_redis())
-    #     for index_class in self.__indexes__ or []:
-    #         if getattr(self, index_class.__key__, None) is None:
-    #             continue
-    #         index = index_class.create_from_model_class(self)
-    #         yield index.save_index(pipe)
-    #
-    #     # remove duplicate on index queue list
-    #     for index, exist_on_index in list_index_map.items():
-    #         if exist_on_index is True:
-    #             model_key_value = getattr(self, self.__key__)
-    #             yield index.remove_from_list(pipe, model_key_value)
-    #
-    #     yield pipe.commit()